Reverse String using StringBuffer class in java
The java.lang.StringBuffer class is a thread-safe, mutable sequence of characters. Following are the important points about StringBuffer
A string buffer is like a String, but can be modified.
It contains some particular sequence of characters, but the length and content of the sequence can be changed through certain method calls.
They are safe for use by multiple threads.
Every string buffer has a capacity.